new includes at libft.h and return value of ft_strlen.c

+++ libft.h
@@ -6,18 +6,21 @@
 /*   By: mpovill- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/18 15:21:40 by mpovill-          #+#    #+#             */
-/*   Updated: 2023/08/18 15:25:49 by mpovill-         ###   ########.fr       */
+/*   Updated: 2023/08/18 16:02:40 by mpovill-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-int	ft_isalpha(int c);
-int	ft_isdigit(int c);
-int	ft_isalnum(int c);
-int	ft_isascii(int c);
-int	ft_isprint(int c);
-int	ft_strlen(char *str);
+# include <stdlib.h>
+# include <unistd.h>
+
+int		ft_isalpha(int c);
+int		ft_isdigit(int c);
+int		ft_isalnum(int c);
+int		ft_isascii(int c);
+int		ft_isprint(int c);
+size_t	ft_strlen(char *str);
 
 #endif
